refactor: move use-locale to locales dir
