+++ app/components/chart-areas.tsx
@@ -7,7 +7,7 @@ import {
   ComponentFieldsFragment,
   useDataCubeObservationsQuery,
 } from "../graphql/query-hooks";
-import { useLocale } from "../lib/use-locale";
+import { useLocale } from "../locales/use-locale";
 import { A11yTable } from "./a11y-table";
 import { Ruler } from "./charts-generic/interaction/ruler";
 import { Tooltip } from "./charts-generic/interaction/tooltip";
