+++ app/components/chart-footnotes.tsx
@@ -3,7 +3,7 @@ import { Box, Link, Text } from "@theme-ui/components";
 import { Fragment, useEffect, useState } from "react";
 import { ChartConfig } from "../configurator";
 import { useDataCubeMetadataWithComponentValuesQuery } from "../graphql/query-hooks";
-import { useLocale } from "../lib/use-locale";
+import { useLocale } from "../locales/use-locale";
 import { DataDownload } from "./data-download";
 
 export const ChartFootnotes = ({
