+++ app/components/chart-scatterplot.tsx
@@ -8,7 +8,7 @@ import {
   ComponentFieldsFragment,
   useDataCubeObservationsQuery,
 } from "../graphql/query-hooks";
-import { useLocale } from "../lib/use-locale";
+import { useLocale } from "../locales/use-locale";
 import { A11yTable } from "./a11y-table";
 import { Tooltip } from "./charts-generic/interaction/tooltip";
 import {
