+++ app/components/dataset-preview.tsx
@@ -4,7 +4,7 @@ import { Loading } from "./hint";
 import { DataTable } from "./datatable";
 import { Trans } from "@lingui/macro";
 import { useDataCubePreviewQuery } from "../graphql/query-hooks";
-import { useLocale } from "../lib/use-locale";
+import { useLocale } from "../locales/use-locale";
 
 export interface Preview {
   iri: string;
