+++ app/components/filters.tsx
@@ -3,7 +3,7 @@ import { Box, Button } from "@theme-ui/components";
 import { useCallback } from "react";
 import { getFilterValue, useConfiguratorState } from "../configurator";
 import { useDimensionValuesQuery } from "../graphql/query-hooks";
-import { useLocale } from "../lib/use-locale";
+import { useLocale } from "../locales/use-locale";
 import { MultiFilterField, SingleFilterField } from "./field";
 import { Loading } from "./hint";
