+++ app/components/links.tsx
@@ -4,7 +4,7 @@ import { useRouter } from "next/router";
 import * as qs from "querystring";
 import { ReactNode } from "react";
 import { Icon, IconName } from "../icons";
-import { useLocale } from "../lib/use-locale";
+import { useLocale } from "../locales/use-locale";
 const createDynamicRouteProps = ({
   pathname,
   query,
